# cfg/paperlib_0.2.1_swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.2.1",
    "title" : "Paper Library Web API",
    "description" : "Paper Library Web API demonstrates integration of swagger with asp.net core + mvc + entity framework"
  },
  "host" : "localhost:4999",
  "basePath" : "/paperlib",
  "schemes" : [ "http" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "tags" : [ {
    "name" : "users",
    "description" : "Operations with users"
  }, {
    "name" : "books",
    "description" : "Operations with books"
  } ],
  "paths" : {
    "/users" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Get all users",
        "description" : "Operation for getting all users",
        "operationId" : "getUsers",
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/user"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "users" ],
        "summary" : "Create new user",
        "description" : "Operation for creating new user",
        "operationId" : "createUser",
        "parameters" : [ {
          "$ref" : "#/parameters/userName"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/users/{id}" : {
      "get" : {
        "tags" : [ "users" ],
        "summary" : "Get an user",
        "description" : "Operation for getting an user by its id",
        "operationId" : "getUser",
        "parameters" : [ {
          "$ref" : "#/parameters/userId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/user"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "users" ],
        "summary" : "Update an user",
        "description" : "Operation for updating an user",
        "operationId" : "updateUser",
        "parameters" : [ {
          "$ref" : "#/parameters/userId"
        }, {
          "$ref" : "#/parameters/userName"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "users" ],
        "summary" : "Delete an user",
        "description" : "Operation for deletting an user by its id",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "$ref" : "#/parameters/userId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/books" : {
      "get" : {
        "tags" : [ "books" ],
        "summary" : "Get all books",
        "description" : "Operation for getting all books",
        "operationId" : "getBooks",
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/book"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "books" ],
        "summary" : "Create new book",
        "description" : "Operation for creating new book",
        "operationId" : "createBook",
        "parameters" : [ {
          "$ref" : "#/parameters/bookName"
        }, {
          "$ref" : "#/parameters/bookOwnerId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    },
    "/books/{id}" : {
      "get" : {
        "tags" : [ "books" ],
        "summary" : "Get a book",
        "description" : "Operation for getting a book by its id",
        "operationId" : "getBook",
        "parameters" : [ {
          "$ref" : "#/parameters/bookId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/book"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "books" ],
        "summary" : "Update a book",
        "description" : "Operation for updating a book",
        "operationId" : "updateBook",
        "parameters" : [ {
          "$ref" : "#/parameters/bookId"
        }, {
          "$ref" : "#/parameters/bookName"
        }, {
          "$ref" : "#/parameters/bookOwnerId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "books" ],
        "summary" : "Delete a book",
        "description" : "Operation for deletting a book by its id",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "$ref" : "#/parameters/bookId"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "type" : "boolean"
            }
          }
        }
      }
    }
  },
  "parameters" : {
    "userId" : {
      "name" : "id",
      "in" : "path",
      "description" : "A string representing the id of an user",
      "required" : true,
      "type" : "integer"
    },
    "userName" : {
      "in" : "query",
      "name" : "name",
      "description" : "A string represents user name",
      "type" : "string"
    },
    "bookId" : {
      "name" : "id",
      "in" : "path",
      "description" : "A string representing book id",
      "required" : true,
      "type" : "integer"
    },
    "bookName" : {
      "in" : "query",
      "name" : "name",
      "description" : "A string represents book name",
      "type" : "string"
    },
    "bookOwnerId" : {
      "in" : "query",
      "name" : "ownerId",
      "description" : "A string represents book owner id",
      "type" : "integer"
    }
  },
  "definitions" : {
    "user" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "name" : {
          "type" : "string"
        },
        "roleId" : {
          "type" : "integer"
        },
        "role" : {
          "$ref" : "#/definitions/role"
        },
        "books" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/book"
          }
        },
        "rentedBooks" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/book"
          }
        }
      }
    },
    "role" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "name" : {
          "type" : "string"
        }
      }
    },
    "book" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer"
        },
        "name" : {
          "type" : "string"
        },
        "ownerId" : {
          "type" : "integer"
        },
        "owner" : {
          "$ref" : "#/definitions/user"
        },
        "readerId" : {
          "type" : "integer"
        },
        "reader" : {
          "$ref" : "#/definitions/user"
        }
      }
    }
  }
}